removes the use of signalfd
Ignore-this: 5241ddd3cb7b498744d2cd70c6da1d95

It is too much complicated to detect when it is really available and
working, and is linux-specific. We use now the classical hack which
consist on writing to a pipe when a signal come.

darcs-hash:20091117175015-c41ad-29dc525d8d7761583163a309a95413ad65fc3e24