support-getpeereid
Ignore-this: bf5116b2268641145ca10008b2ff81df

Add support for Lwt_unix.get_credentials on MacOS X/OpenBSD.  A peer uid/gid
can be obtained via getpeereid(2) on some platforms, but the peer pid will not
be available. In this case, fill in the uid,gid in the credentials, and set pid
to -1 to signify that it is not available.

darcs-hash:20120322113842-50edc-95dc7d67eee7fcf7d1453b9edf3546f6a9f0b81c