add is_closed to Lwt_stream.t
Before this commit the Lwt_stream API provided no reliable way to detect
if a stream had been closed. While on_termination notifies the user if
the stream is closed in the future, it does not tell the user if the
stream was closed in the past.

Introducing the is_closed function addresses this problem, allowing the
user to detect if a stream has been closed at any point in time.