Add closed, is_closed, and improve on_termination.
Before this commit, the Lwt_stream API provided no reliable way to
detect if a stream had been closed. While on_termination notified the
user if the stream was closed in the future, it did not tell the user
if the stream had been closed in the past.

The is_closed function addresses this problem, allowing the user to
detect if a stream has already been closed at any point in time.

The closed function returns a thread that will sleep until the stream
is closed.

Lwt_stream.on_termination is now implemented in terms of Lwt.closed.

Lwt_stream.on_termination will be deprecated and ultimately removed in
favor of Lwt_stream.closed.