implement Lwt_stream.on_termination in terms of Lwt.closed
As per the discussion on the issue tracker, Lwt_stream.on_termination
will be deprecated and ultimately removed in favor for
Lwt_stream.closed. To facilitate that, this commit introduces a new
implementation of Lwt_stream.closed that does not rely on termination
hooks, and implementation Lwt_stream.on_termination in terms of
Lwt_Stream.closed.