change close semantics for fixed-length sources
Previously, the of_list, of_array, and of_string constructors each would
produce a stream s for which

  is_closed s = false

This is counter-intuitive as all the elements of the stream s have been
generated and should be consumable without blocking. This commit changes
the behavior of this constructors such that

  is_closed s = true

immediately upon creation.