Fix deadlock in Lwt_main.at_exit
The underlying problem is an interaction between calling Lwt_main.run
from a thread, and deferring a thread wakeup using Lwt.wakeup_later.
The example is based on Lwt_io.flush, as in #48:

Thread A:
- calls flush on channel C,
- blocks on I/O.

Thread B:
- is running inside a wakeup,
- calls Pervasives.exit,
- which eventually calls Lwt_main.run in order to run handlers
  registered with Lwt_main.at_exit,
- which creates thread D to flush channel C,
- thread D blocks in the operations queue of channel C, waiting for A,
- thread B continues running Lwt_main, waiting for D and performing I/O,
- which eventually wakes thread A.

Thread A:
- sees thread D in the queue, and calls Lwt.wakeup_later on it.

At this point, thread D will run the next time the outermost wakeup
(thread B's) completes. However, that wakeup is running Lwt_main.run,
which is waiting, synchronously, for thread D to complete. This is the
deadlock.

This patch is really a band-aid. In the long term, Lwt likely needs a
more principled approach to cleanup on exit.

Note that the docs for Lwt_main.run say that it shouldn't be called from
inside threads, as done above in thread B. This is a programming error
inside Lwt, but it's not very clear how to avoid it.

Fixes #48.
Fixes #114.