test: lwt_mutex: Make sure all threads have ended
While integrating 9cbce8 into our testsuite (lwt 2.5.0 based) errors were
observed where the mutex was still locked.

The difference is that in 2.3.2 the thread is cancelled, yielding a failure
but in 2.5.0 the cancel becomes a nop and the task still needs to
run. Depending on the circumstances this hasn't happened before the end
of the test, even though the mutex is still taken.

Therefore, ensure the task is finished, taking the difference in
cancel-behavior of 2.3.2 and 2.5.0 into account (should anyone want to
backport or run on an older version)