test_lwt_mutex: cover PR202 as requested in PR261
The original issue was present in 2.3 series of lwt, where
canceling a thread immediately after it unlocking a mutex
could thwart the unlock, resulting in a forever-locked mutex.
See https://github.com/ocsigen/lwt/pull/202 for a fix on 2.3.2.

Subsequent changes in lwt core fixed this issue. A reproducing
test was submitted at https://github.com/ocsigen/lwt/pull/261, but
as reported there the test changed in this patch is a superset
under the condition that 'wakeup' was changed in 'wakeup_later'.

Maintainer:
Submitted in #262.
Resolves #262.
Resolves #261.