Tests: add coverage analysis for OCaml code
Install Bisect_ppx:

  opam install bisect_ppx

Enable with:

  ocaml setup.ml -configure --enable-tests --enable-coverage --disable-camlp4

Then run

  make coverage

This will build and run the tests, and generate a coverage report in the
_coverage/ subdirectory (view _coverage/index.html).

Building coverage with Camlp4 enabled results in an assertion failure in
the type checker.

The coverage number reported after this commit is not particularly
accurate: C code is not included, and some OCaml files are not linked
into the testing binaries at all. These problems will be addressed in
future commits, as necessary, and definitely before coverage numbers are
"touted."