AppVeyor: temporarily disable Cygwin+libev build
It is producing failures such as

  https://ci.appveyor.com/project/aantron/lwt/build/342/job/uui64qa8lams04vo#L931

and I don't currently have time to debug it. The purpose of this commit
is to decrease noise for contributors. The libev build should generally
succeed when the non-libev build succeeds anyway, i.e. the chance of
introducing a failure that would be detected only by the libev build is
fairly low.