lwt.ml: one function for creating pending promises
In the final state of lwt.ml, there is only one function that creates
pending promises (i.e., applies the Pending constructor). This commit
approaches that state, but add_task_r and add_task_l still apply
Pending themselves internally. That is best refactored after the future
commit that converts promise state to a GADT.

Decreasing the number of ways to create a pending promise now simplifies
the diff of the next commit.