lwt.ml: make promise state into a GADT
The GADT has two additional type parameters, and these encode the state
of each promise in its type. For example, if one has a reference to a
promise of type

  ('a, underlying, pending) promise

then one knows that the only possible state the promise can be in is
Pending. This is in contrast to a reference to the same promise with
type 'a Lwt. With such a reference, one does not know what state the
promise is in -- it could be Resolved, Failed, Pending, or Unified_with.

This representation eliminates a large number of assert false cases.

On the other hand, this representation can introduce a problem: if the
Lwt code holds a reference to an ('a, underlying, pending) promise, and
also hands it out to the user, the user might somehow change the state
of the promise -- and this state change is not reflected in the type of
the reference held by Lwt. This dangerous situation is typical when a
promise is created by a function, such as bind, and also used in a
callback registered by that function to run later.

This complication is addressed by introducing an explicit cast

  let State_may_now_be_pending_proxy p = may_now_be_proxy p in ...

This is the first thing done by the callback. It is a no-op, whose only
purpose is to shadow p with a new reference of type

  ('a, $Unknown, pending) promise

which reflects that p may have become a proxy, in the time until the
callback was called. Similarly, when Lwt itself mutates the state of a
promise, it does so with

  let State_may_have_changed p = set_promise_state p state in ...

This shadows p with a reference that inherits type variables from
state.

When Lwt receives a totally arbitrary promise from the user, the no-op
cast

  let Internal p = to_internal_promise p in ...

shadows the original p, of type 'a Lwt.t, with another reference of type

  ('a, $Unknown_1, $Unknown_2) promise

Unlike the other two casts, this one is almost impossible to
accidentally omit, as almost none of the internal helpers would accept
an ordinary 'a Lwt.t.

These casts are an annoying side effect of encoding mutable state in the
type parameters of a GADT. At the same time, they explicitly document
all possible state changes for readers of the Lwt code.