Lwt_glib: fix Homebrew build
An upstream update of Homebrew's glib package causes pkg-config to emit
arguments that look like this:

  -Wl,-framework -Wl,CoreFoundation

This is syntax for options that the C compiler is to pass to the linker.
However, ocamlmktop expects these options in this syntax:

  -framework CoreFoundation

The code added by this commit searches pkg-config output for arguments
of the first form, and replaces them with arguments of the second form.

[skip appveyor]