Remove dependancy on Base and Stdio
mainly because they are not available on 4.02.0.  It means we somewhat
hack generation of the sexps used to communicate C flags.

Also, dont pass ocamlc to the unix discover script.  The default uses
"ocamlfind ocamlc", which makes us depend on ocamlfind, but this
seems to be needed due to the way options are passed to ocamlc
(-ccopt -o -ccopt <prog>) which doesnt work on ocaml < 4.04
(without ocamlfind).