address jbuilder code review issues
1. remove unused files; TODO, _opam, bad_packages

2. fix library description (thread->promise) in discover.ml

3. update .gitignore

4. remove jbuilder line in readme

5. update opam file post install message and remove
   ocaml-migrate-parsetree dep, revert post install message

6. get build to work without presence of camlp4

according to janestreet/jbuilder#51 we need to move the 3 camlp4
packages into seperate directories.

I note a general opam packing issue where if a system installed
camlp4 is in the PATH when ocamlfind is installed, it adds a META
file which pretends those packages are available and compatible with
any current OPAM switch, which it is not!

I think this is an upstream camlp4 packaging issue

7. Add a custom META file

This is only to support the correct syntax predicates for camlp4
and ocamlfind.

Also requires various exists_if clauses