Remove dummy ppx_lwt.ml
The commit after this one is the port of Lwt to jbuilder. As part of
that port, ppx_lwt_ex.ml is renamed to ppx_lwt.ml. The original, dummy
ppx_lwt.ml is deleted.

If this is all done in one commit, the resulting blame shows that
ppx_lwt.ml was entirely written during the jbuilder port. This is
misleading, and would harm history investigations in the future.

For this reason, the deletion of the dummy ppx_lwt.ml is factored out
into its own commit.

In this commit, the PPX does not build, nor pass tests.

Maintainer's note: this actual commit was created by Anton Bachin
(@aantron); though the change itself was originally made by the author,
Andrew Ray (@andrewray).

[skip ci]