Rename META: it is not a template at the moment
Jbuilder META templates include the line

  JBUILDER_GEN

(leading with a pound sign, which I have omitted from the commit message
to prevent git treating it as a comment).

We have, for now, inlined all Jbuilder-generated content into Lwt's META
file, and modified it. We will probably let Jbuilder do all the work in
a future release that updates the packaging, perhaps Lwt 4.0.0.