Lwt_list: add [@ocaml.tailcall] in some functions
The annotation is added to functions that were modified in #347. It will
be added to other functions in Lwt_list during a future thorough review
of that module.

[@ocaml.tailcall] was introduced in OCaml 4.03.0. On 4.02, the
annotation is silently ignored, so it is safe. Working on 4.03.0 and
higher is enough for any problems with tail recursion to be caught in
CI.