Coverage for more files with Bisect_ppx 1.3.0
Bisect_ppx 1.3.0 adds the -conditional workaround, which makes it
somewhat easier to use with Jbuilder (it's no longer necessary to write
jbuild files in OCaml syntax). This commit extends coverage to more Lwt
files, based on this workaround.

Generating coverage analysis now requires OPAM package react to be
installed, because coverage is generated for lwt_react.