Begin factoring out PPX to its own package
This commit creates a new opam package lwt_ppx, and a new Findlib
package of the same name. This Findlib package will replace lwt.ppx,
from existing opam package lwt.

For now, lwt_ppx is just an alias for lwt.ppx, so that users have
several months to gently switch over.

There is one defect: with ocamlfind and Ocamlbuild, PPX options have to
be passed to lwt.ppx, not lwt_ppx. When the switchover is complete,
uses will suddenly have to pass options to lwt_ppx. This is not a
problem when using Jbuilder and/or a PPX driver.