Tests: call Lwt_main.run in only one place
This is a step towards porting the test suite to run with JS promises.
On Node (and in browsers), there is no equivalent to Lwt_main.run. The
JS engine handles the I/O loop internally instead. So, Lwt_main.run is
not available as a way to sychronously wait for asynchronous code to
finish, nor is there any other apparent way to do this.

This commit makes the test framework almost entirely asynchronous.

A few tests had to be adjusted, because tests are now run in callbacks
triggered, sometimes, by Lwt.wakeup_later. Nested calls to
Lwt.wakeup_later are deferred, so some synchronous promise state checks
began to fail when test cases depended on Lwt.wakeup_later to update
promise state.

See also

  https://github.com/ocsigen/lwt/issues/428#issuecomment-340012047