Lwt_unix: keep thread count accurate
See #493.

Before this commit, Lwt performed the following steps when it needed to
start a worker thread, while holding its thread pool mutex:

1. Increment Lwt's internally-tracked worker thread count.
2. Release the thread pool mutex.
3. Call a helper (launch_thread) that calls pthread_create.

If pthread_create failed, launch_thread would raise an OCaml exception,
by calling unix_error. The result was that if pthread_create failed,
Lwt's thread count would diverge from the real thread count, since the
exception is not handled, and the thread count is not decremented.

After this commit,

1. launch_thread is called first, and calls pthread_create. It now
   returns a result code instead of raising an exception. This makes it
   possible to run launch_thread with the pool mutex still held.
2. If launch_thread succeded, the thread count is incremented.
3. The mutex is released.
4. If launch_thread failed, Lwt calls unix_error to raise an exception.

The Windows version of launch_thread does not seem to do any error
checking, so this commit always has it return the success code (0). This
is probably incorrect, but it is not more incorrect than before.

launch_thread is also used with the ASYNC_SWITCH job execution scheme,
so its code has been adjusted slightly. The ASYNC_SWITCH scheme was
experimental, is now effectively deprecated, and will probably be
removed completely in the future.