Better custom configuration (#693)
I was tripped up by the example in the travis script when trying to
build lwt with libev without opam.
`dune exec src/unix/configurator.exe -- -use-libev true` doesn't really
do anything since the `lwt_config` file will be generated at the project root
instead of where dune expects it.

Also update the `dune` file to allow proper fallback. Currently dune
gives a warning when a file is both in the source tree and generated by a rule.
This will be an error in future versions of dune.