Lwt.get: don't use exceptions for control flow
Another option for this change was to use Map.S.find_opt. However, this
requires OCaml 4.05, but Lwt still supports 4.02. The code in this
commit avoids having to select different implementations depending on
the OCaml version. This code is not performance-critical, so doing two
descents into the map should not be a problem.

Part of #714.