Use accept4 to avoid 2 fcntl system calls
accept4 accepts flags to tell to create the file descriptor in
non blocking mode. This way we don't need to set it later with
Unix.set_nonblock.

Also, since we know that all the calls to accept will be on
non blocking sockets we can avoid releasing the runtime lock.