Skip "bind: Unix domain" on Hurd (#816)
At the moment the Hurd cannot provide the name (i.e. the path) of a
connected Unix socket, and thus getsockname always returns an empty
string. Hence skip this test on Hurd for now, as the path comparison
does not work, while the test (and the APIs underneath) works fine
otherwise.

Fixes #802