Add PF_INET6 to socket_domain_table on Windows
Lwt provides an emulation of socketpair on Windows, using PF_INET6
sockets. It has to duplicate the table of Unix.socket_domain. The type
reads as follows:

    type socket_domain =
    | PF_UNIX          (* Unix domain *)
    | PF_INET          (* Internet domain (IPv4) *)
    | PF_INET6         (* Internet domain (IPv6) *)

However, the Lwt `socket_domain_table` has only two entries. If
`socketpair` with PF_INET6 sockets is called, then a value
out-of-bounds of the array will be requested.

Fixed by adding PF_INET6 to the array, thus having the same arity as
the type it's backing.