Use is_blocking in dup and dup2 to fix ENOTSOCK on Windows
On Windows, calling Unix.clear_non_block and Unix.set_non_block is
only supported on socket file descriptors. The function is_blocking
already deals with the corner cases, but the current implementations
of dup and dup2 don't make use of it. Thus, when then dup or dup2
wrappers are called with file socket descriptors, ENOTSOCK errors are
raised. The fix is to reuse is_blocking in the wrappers.

Signed-off-by: Antonin Décimo <antonin@tarides.com>