Lwt_react: no longer compare structurally the arguments of S.l[2-6]_s
This avoids a failure when some inputs cannot be compared structurally
(if they contain functional values, for instance).

Internally, Lwt_react.S.l2 is implemented using React.S.l2 and
Lwt_react.S.map_s. We were not providing an equality function to
React.S.l2, so the default structural equality was used. Now, we use the
function (fun _ _ -> false). So, the function parameter is applied each
time any of the input change, which is exactly what we would expect if
we implemented Lwt_react.S.map_s directly.