Set all pipes to cloexec
This is done to prevent possible race conditions, and for consistency.
The spawn function is now responsible for ensuring that file
descriptors can be inherited by subprocesses (this is always the case
on unix when using dup2, but can be trickier on Windows).

We can also remove to_close file descriptors in Lwt_process spawn,
because the file descriptors are the parent's end of the pipes which
must be closed in the child; however switching the pipes to cloexec
means they won't be inherited by the child at all.