Some _SC* macros are not available from the Cosmopolitan libc
This patch is probably the more invasive about Esperanto/Cosmopolitan. It adds
a new compilation path when [__ESPERANTO__] is defined. A note was added to
explain where it comes from: the cross-compiler [arch-esperanto-none-static-cc]
systematically defines [-D__ESPERANTO__] in order to let third-party libraries
(such as lwt) to choose different compilation paths. And here we are!

Cosmopolitan does not defines required [_SC*] constants and
[unix_get_pw_gr_nam_id_job] can not be compiled. We discard this piece of code
when we want to compile with Cosmopolitan as it's the case for Android too.