Change wrong O_CLOEXEC to O_KEEPEXEC on null fd for sub-processes
On Unix O_CLOEXEC on a file descriptor was introduced in
9de01c9f8ddb5f4956f1473b657c055c1974642f. It is superfluous and buggy:
we're already in the forked process at this point! So setting it
cloexec between the actual fork and exec will close the file
descriptor in the sub-process, making all read and writes to it fail.

On Windows, it is wrong as well: we want the child process to inherit
the NUL file descriptor, so it mustn't be cloexec, and it must be
closed after the spawn in the parent process.

It wasn't caught in the tests because they weren't complete enough,
fixed in the parent commit.

The files descriptors must remain keep-on-exec, so we set this file
instead explicitly.