Fix signal masking race in worker_loop
If an OCaml signal handler runs in a Lwt worker thread then the process
will dereference a NULL pointer and crash. The worker tried to avoid
this by masking out all signals after starting, but by then it may be
too late. Now, we mask out signals before creating a worker thread and
then restore the mask in the main thread afterwards.