core/lwt.ml: box resolved states
This commit guarantees that, even if we later weaken the type
information available on `state`, we keep precise guarantees on
`result` values -- they are not a subset of possible states anymore,
but a separate type.

There may be a performance impact to this change, so running
benchmarks would be useful.